pyproject.toml
[tool.poetry]
name = "embedded-utils"
version = "0.1.0"
description = ""
authors = ["Guilhermwn <limaoplaykglp@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
embedded-utils = "embedded_utils.main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = "0.12.3"
click = "8.1.7"
colorama = "0.4.6"
lorem = "0.1.1"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
pygments = "2.18.0"
pyserial = "3.5"
rich = "13.7.1"
shellingham = "1.5.4"
tabulate = "0.9.0"
typing-extensions = "4.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
